Extract the standalone Geam CLI
- Move standalone command ownership into geam-cli.
- Keep the root geam binary as a thin installation target.
- Preserve generated runner source and command behavior.
