+++ Cargo.toml
@@ -19,8 +19,22 @@ include = [
 ]
 
 [workspace]
-members = [".", "core", "builtins/stdlib", "builtins/json", "builtins/time"]
-default-members = [".", "core", "builtins/stdlib", "builtins/json", "builtins/time"]
+members = [
+    ".",
+    "core",
+    "cli",
+    "builtins/stdlib",
+    "builtins/json",
+    "builtins/time",
+]
+default-members = [
+    ".",
+    "core",
+    "cli",
+    "builtins/stdlib",
+    "builtins/json",
+    "builtins/time",
+]
 resolver = "3"
 
 [workspace.package]
@@ -41,6 +55,7 @@ clap = { version = "4", features = ["derive"] }
 ecow = "=0.2.6"
 flate2 = "1"
 geam-core = { path = "core", version = "=0.1.2" }
+geam-cli = { path = "cli", version = "=0.1.2" }
 geam-json = { path = "builtins/json", version = "=0.1.2" }
 geam-stdlib = { path = "builtins/stdlib", version = "=0.1.2" }
 geam-time = { path = "builtins/time", version = "=0.1.2" }
@@ -70,37 +85,16 @@ ureq = { version = "3", features = ["json"] }
 vec1 = "1"
 
 [dependencies]
-base64.workspace = true
-bitvec.workspace = true
-camino.workspace = true
-cargo_metadata.workspace = true
-clap.workspace = true
-ecow.workspace = true
-flate2.workspace = true
+geam-cli.workspace = true
 geam-core.workspace = true
 geam-json.workspace = true
 geam-stdlib.workspace = true
 geam-time.workspace = true
-gleam-core.workspace = true
-half.workspace = true
-hex.workspace = true
-hexpm.workspace = true
-im.workspace = true
-miette.workspace = true
-num-bigint.workspace = true
-num-traits.workspace = true
-rand.workspace = true
-semver.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-sha2.workspace = true
-tar.workspace = true
-tempfile.workspace = true
-thiserror.workspace = true
-toml.workspace = true
-unicode-segmentation.workspace = true
-ureq.workspace = true
-vec1.workspace = true
 
 [dev-dependencies]
+camino.workspace = true
+ecow.workspace = true
+gleam-core.workspace = true
 miette = { workspace = true, features = ["fancy-no-backtrace"] }
+tempfile.workspace = true
+toml.workspace = true
