+++ cli/src/command.rs
