+++ cli/src/provider/manifest.rs
