+++ cli/src/provider/metadata.rs
