+++ cli/src/provider/registry/archive.rs
