+++ cli/src/provider/registry/configuration.rs
