+++ cli/src/provider/registry/discovery.rs
