+++ cli/src/provider/registry/index.rs
