+++ cli/src/provider/registry/search.rs
