+++ cli/src/runner.rs
@@ -1,8 +1,5 @@
-#[path = "runner/cargo.rs"]
 mod cargo;
-#[path = "runner/generator.rs"]
 mod generator;
-#[path = "runner/source.rs"]
 mod source;
 
 pub(super) use cargo::{CargoLock, RunnerChecker, RunnerExecutor, SystemCargo, reconcile_lock};
